mysite/requirements.txt
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.3
Django==5.2.8
django-allauth==65.13.0
django-cors-headers==4.9.0
django-filter==25.2
django-modeltranslation==0.19.17
django-phonenumber-field==8.3.0
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
googletrans==4.0.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
openapi-codec==1.3.2
packaging==25.0
phonenumbers==9.0.18
phonenumberslite==9.0.18
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
simplejson==3.20.2
sniffio==1.3.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
